telegram_utils/requirements.txt
python-telegram-bot>=22.5
mcrcon>=0.7.0
PyYAML>=6.0.3